e2fsck: Only release clusters when shortening a directory during a rehash
When the rehash process is running on a bigalloc filesystem, it compresses all
the directory entries and hash structures into the beginning of the directory
file and then uses block_iterate3() to free the blocks off the end of the file.
It seems to call ext2fs_block_alloc_stats2() for every block in a cluster,
which is unfortunate because this function allocates and frees entire clusters
(and updates the summary counts accordingly).  In this case e2fsck writes out
incorrect summary counts.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>